+++ components/CHANGELOG.md
@@ -10,6 +10,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 ### Changed
+- Select API changed, given the following options:
+
+    ```js
+    options = [
+      {label: "label1", value: "value1},
+      {label: "label2", value: "value2},
+    ]
+    ```
+
+    Old API:
+
+    ```js
+    <Select options={options} value={options[1]}>
+    ```
+
+    New API:
+
+    ```js
+    <Select options={options} value="value2">
+    ```
 
 ### Deprecated